pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "styrecon"
version = "0.1.0"
description = "StyRecon — Pura Vida Ops (scope-aware, stateful bug bounty recon orchestrator)"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "vdmarcoss" }]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "PyYAML>=6.0.1",
  "sqlmodel>=0.0.22",  # not required in v0.1 runtime path, kept for future evolution
]

[project.scripts]
styrecon = "styrecon.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/styrecon"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]